Match widget fixes
This change includes two changes to the match widgets. First, match
widgets no longer contract and expand. Second, the styles for the match
widgets has been adjusted to increase the size of the W/L and make the
@/vs bold, as well as increasing the spacing above the venue.

This change also includes a tweak to the display for the error page.