Make header text size consistent and fix mobile style issues
This change makes the header text size a fixed size in px rather than em
to ensure that the size is consistent for pages and posts.

This change also ensures that the secondary bar appears after the main
bar and not behind it when the screen size gets small, as well as
removing a border from the bottom navigation section that was very odd
looking on mobile devices.